refactor(app_view): split editor form into editor.rs
Lifts the profile-editor form rendering + state-bridging out of
mod.rs into a new `app_view::editor` child module. About 1,400
lines covering:

  * `build_editor` / `apply_editor_to_profile` — the bidirectional
    bridge between a `Profile` and the live form widget state.
  * `editor_fields_match` — dirty-check predicate that drives the
    Save button's brightness.
  * `EditorRender` + its `from(&Editor, &Context)` constructor —
    the cloned-out widget handles a render pass needs, so the
    render code can hand `cx: &mut Context<AppView>` to form
    helpers without keeping `&self.editor` borrowed alongside.
  * `form_pane` — the public entry; called from `Render for AppView`.
  * `form_header`, `form_body`, `form_tab_nav`, `section_card`,
    `section_card_with_desc`, `labeled`, `bool_field`,
    `bool_field_hinted`, `connection_card`, `terminal_card`,
    `highlighting_pane`, `advanced_pane`, `theme_card`,
    `control_lines_card`, `output_card`, `paste_safety_card` —
    the internal helpers that compose the form's visual tree.
  * `detect_missing_drivers` + `driver_banner_row` — the
    unenrolled-USB-driver warning shown above the Serial Port
    field; only the form renders this so it lives here.

Visibility plan: `pub(super)` on the four items mod.rs actually
references (`build_editor`, `apply_editor_to_profile`, `form_pane`,
`struct EditorRender` + `EditorRender::from`). The rest stay
private — `form_header` / `form_body` / etc. are only ever called
from within editor.rs.

`struct Editor` + `enum EditorTab` stay in mod.rs because the
parent's methods on `AppView` (set_editor_tab, save_editor,
toggle_local_echo, …) do direct field access on Editor, and the
descendants-see-parent-private rule means editor.rs can still
read those fields without per-field `pub(super)` widening. The
split is along "form rendering & state-bridging logic" lines
rather than "Editor data type and friends."

mod.rs drops by ~1,395 lines (5,651 → 4,256). editor.rs lands at
1,441 lines — large, but every item is part of the editor form
pipeline.

Verified: `cargo clippy --all-targets -- -D warnings` and
`cargo test` both pass (63 tests, unchanged).
